refactor: simplify FaceMirror orchestrator and modularize face landmarker demo components
