+++ demos/face_mirror/index.html
@@ -40,6 +40,7 @@
   </head>
 
   <body>
+    <face-mirror-status></face-mirror-status>
     <script type="module" src="main.js"></script>
   </body>
 </html>
